Add files via upload
In this Project my groupmate and I choose to explore how social applications such as Facebook, Instagram, twitter, YouTube, and so on have expanded in 3 major sections: Before covid-19 Pandemic, During the Pandemic, and near the "end" of the Pandemic. Unfortunately, we where not able to access such data freely online and adjusted our goal to exploring internet usage on a global scale. Comparing data provided from site such as ITU, Pew Research Center, The World Bank, Our World in Data, and statcounter, we found some intersting results in how the US is ranked with other countries around the globe in terms of internet usage using broadband or mobile/wireless sources.  With our limited access we also explored how certain social media applications usage have evolved within a time span of 10 years, 2008 to 2018. The final discover we found intersting in our research/project was an increase need of desktops compared to mobile/laptops since the start of the pandemic.